+++ src/main/java/it/gesev/dao/TestataMovimentoDAOImpl.java
@@ -41,7 +41,7 @@ public class TestataMovimentoDAOImpl implements TestataMovimentoDAO {
 	{
 		logger.info("Ricerca testata movimento con ID fornitore " + idFornitore);
 		
-		List<TestataMovimento> testataMovimento = testataMovimentoRepository.getTestataMovimentoByIdFornitore(idFornitore);
+		List<TestataMovimento> testataMovimento = testataMovimentoRepository.getTestataMovimentoByIdFornitore(idFornitore.intValue());
 		if(testataMovimento.size() > 0)
 			return testataMovimento;
