+++ src/main/java/it/gesev/repository/DerrataRepository.java
@@ -8,13 +8,13 @@ import org.springframework.data.jpa.repository.Query;
 
 import it.gesev.entities.Derrata;
 
-public interface DerrataRepository extends JpaRepository<Derrata, Long>
+public interface DerrataRepository extends JpaRepository<Derrata, Integer>
 {
 	public Optional<Derrata> findByDescrizioneDerrata(String descrizioneDerrata);
 	
 	
 	@Query("select d from Derrata d where d.tipoDerrata.codice = :tipoDerrataId")
-	public List<Derrata> findAllByDerrataId(long tipoDerrataId);
+	public List<Derrata> findAllByDerrataId(int tipoDerrataId);
 	
 	@Query("select max(derrataId) from Derrata")
 	public Integer getMaxDerrataId();
