+++ src/main/java/it/gesev/repository/FornitoreRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import it.gesev.entities.Fornitore;
 
-public interface FornitoreRepository extends JpaRepository<Fornitore, Long> 
+public interface FornitoreRepository extends JpaRepository<Fornitore, Integer> 
 {
 	public Optional<Fornitore> findByDescrizione(String descrizione);
 }
